% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFactorCosts.R
\name{calcFactorCosts}
\alias{calcFactorCosts}
\title{calcFactorCosts}
\usage{
calcFactorCosts(
  datasource = "USDA",
  otherLivst = FALSE,
  subsectors = FALSE,
  unit = "constant 2017 US$MER"
)
}
\arguments{
\item{datasource}{only source available is "USDA" (calculates factor costs by applying factor cost share from USDA
to VoP from FAO)}

\item{otherLivst}{boolean: should FAO livestock categories that can't be matched to MAgPIE categories (i.e. beeswax,
wool, silkworms, and honey) be reported as "livst_other"?}

\item{subsectors}{boolean: should output be aggregated or reported by crop and livestock subsectors?}

\item{unit}{output currency unit based on the toolConvertGDP function from the  GDPuc library}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
combines factor costs for crop and livestock production in mio. US$MER05
}
\examples{
\dontrun{
calcOutput("FactorCosts")
}
}
\author{
Debbora Leip
}
