% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFractionInputsUSDA.R
\name{calcFractionInputsUSDA}
\alias{calcFractionInputsUSDA}
\title{calcFractionInputsUSDA}
\usage{
calcFractionInputsUSDA(
  products = "kcr",
  keepConstantExtrapolation = TRUE,
  interpolate = TRUE
)
}
\arguments{
\item{products}{either "kcr" for crops, or "kli" for livestock}

\item{keepConstantExtrapolation}{boolean: should constant extrapolation from Fuglie et al. be kept?}

\item{interpolate}{boolean: should the data be interpolated to the middle of the decade?}
}
\value{
magpie object of the shares of the factor requirements in agriculture (capital, labor, materials, land).
}
\description{
Calculates the factor factor shares for crop or livestock production from USDA'S Inputs shares.
}
\examples{
\dontrun{
a <- calcOutput("FractionInputsUSDA")
}

}
\seealso{
[calcOutput()]
}
\author{
Edna J. Molina Bacca, Debbora Leip
}
