% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandRent.R
\name{calcLandRent}
\alias{calcLandRent}
\title{calcLandRent}
\usage{
calcLandRent(unit = "constant 2017 US$MER")
}
\arguments{
\item{unit}{output currency unit based on the toolConvertGDP function from the  GDPuc library}
}
\value{
magpie object of the land factor requirements in USD/ha per crop
}
\description{
Calculates factor intensity in crop production for ladn rent from USDA (Inputs share)
and FAO (Value of Production) and productivity in constant 2017 US$MER per ha.
}
\examples{
\dontrun{
a <- calcOutput("LandRent")
}

}
\seealso{
[calcOutput()]
}
\author{
Edna J. Molina Bacca
}
