% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProductionCosts.R
\name{calcProductionCosts}
\alias{calcProductionCosts}
\title{calcProductionCosts}
\usage{
calcProductionCosts(datasource = "Vittis", unit = "constant 2017 US$MER")
}
\arguments{
\item{datasource}{Datasource of production costs, currently only "Vittis"}

\item{unit}{output currency unit based on the toolConvertGDP function from the  GDPuc library}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates agricultural production costs (split into different cost categories)
}
\examples{
\dontrun{
calcOutput("ProductionCosts", source = "Vittis")
}

}
\author{
Debbora Leip
}
