% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRegFactorShare.R
\name{calcRegFactorShare}
\alias{calcRegFactorShare}
\title{calcRegFactorShare}
\usage{
calcRegFactorShare(datasource = "USDA", caseStudies = "CountryCaseStudies")
}
\arguments{
\item{datasource}{Only USDA available}

\item{caseStudies}{The case studies to be used for the regression (either CountryCaseStudies or
CaseStudiesDirectMapping). Default is CountryCaseStudies, as including regional case studies weakens the direct
link to GDP per capita (and results in a regression with non-normally distributed residuals).}
}
\value{
MAgPIE object at global level with slope and intersect as items
}
\description{
This function calculates the regression parameters (a and b) for the function Share=a*log10(GDP)+b
Where share is the adjusted share between capital and labour.
}
\examples{
\dontrun{
calcOutput("calcRegFactorShare")
}

}
\seealso{
[calcOutput()],[calcFactorIntensity()]
}
\author{
Debbora Leip, Edna J. Molina Bacca
}
