% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVoPAFF.R
\name{calcVoPAFF}
\alias{calcVoPAFF}
\title{calcVoPAFF}
\usage{
calcVoPAFF(unit = "constant 2017 US$MER")
}
\arguments{
\item{unit}{output currency unit based on the convertGDP function from the  GDPuc library}
}
\value{
magpie object. in mio. USD
}
\description{
Calculates the overall value of production of the agriculture,
forestry and fisheries sectors. Forestry and Fisheries are calculated from exports values.
}
\examples{
\dontrun{
a <- calcOutput("VoPAFF")
}

}
\seealso{
[calcOutput()]
}
\author{
Edna J. Molina Bacca, Debbora Leip
}
