% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVoPcrops.R
\name{calcVoPcrops}
\alias{calcVoPcrops}
\title{calcVoPcrops}
\usage{
calcVoPcrops(fillGaps = TRUE, unit = "constant 2017 US$MER")
}
\arguments{
\item{fillGaps}{boolean: should gaps be filled using production * prices (where production data is available)?}

\item{unit}{output currency unit based on the toolConvertGDP function from the  GDPuc library}
}
\value{
magpie object. in mio. USD or fraction
}
\description{
Calculates the value of production of individual production items or
its fraction compared to overall Value of Production (Agriculture, Fish, Forestry).
}
\examples{
\dontrun{
a <- calcOutput("VoPcrops")
}

}
\seealso{
[calcOutput()]
}
\author{
Edna J. Molina Bacca, Debbora Leip
}
