% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWeeklyHoursILO.R
\name{calcWeeklyHoursILO}
\alias{calcWeeklyHoursILO}
\title{calcWeeklyHoursILO}
\usage{
calcWeeklyHoursILO(projections = FALSE, dataVersion = "July23")
}
\arguments{
\item{projections}{boolean, should weekly hours be projected (by keeping constant) up to 2150?}

\item{dataVersion}{which version of the ILO input data to use. "" for the oldest version and
old regression, or "monthYear" (e.g. "July23") for newer data}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates complete dataset of mean weekly hours worked by people employed in agriculture, forestry and
fishery based on ILO dataset
}
\examples{
\dontrun{
calcOutput("WeeklyHoursILO")
}
}
\author{
Debbora Leip
}
