% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTFPUSDA.R
\name{correctTFPUSDA}
\alias{correctTFPUSDA}
\title{correctTFPUSDA}
\usage{
correctTFPUSDA(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
corrected mapie object on national-scale costs of production for 10
crops, disaggregated in 9 distinct cost elements
}
\description{
Remove values that are just constant extrpolation
}
\seealso{
[readTFPUSDA()]
}
\author{
Debbora Leip
}
