% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRegressionsILO.R
\name{readRegressionsILO}
\alias{readRegressionsILO}
\title{readRegressionsILO}
\usage{
readRegressionsILO(subtype = "AgEmplShare", version = "July23")
}
\arguments{
\item{subtype}{Type of ILOSTAT data for which regression coefficients should be read
\itemize{
\item `AgEmplShare`: regression coefficients for sqr(ag empl share) ~ log(GDP pc PPP)
\item `HourlyLaborCosts`: regression coefficients for hourly labor costs ~ GDP pc MER (old version) or
log(hourly labor costs) ~ log(GDP pc MER) (new version)
}}

\item{version}{version of regression and underlying data. "" for the oldest version, or "monthYear" (e.g. "July23")
for newer version}
}
\value{
regression coefficients as MAgPIE object
}
\description{
Read regression coefficients which are used to fill missing values of ILO datasets
}
\examples{
\dontrun{
  readSource("RegressionsILO", "AgEmpl")
}
}
\author{
Debbora Leip
}
